Tab Panel Registers on `willInsertElement`
In experimentation, I was able to see that the registration of panels
was happening late. Switching the `registerWithTabs` method trigger
from `didInsertElement` to `willInsertElement` event has solved the
issue for me.

That being said, this is hackish at best. I cannot fully explain the
problem this fix solves.

More investigation is needed as to why the `tab` computed property
is not correctly updated after the `tabs` component's `tabPanels`
property is set/updated. There is probably a more stable fix down
that path.

This should hold over those who need this component for now.